// src/app/models/forecast-to-plot.dto.ts
// Generated by https://quicktype.io

export interface ForecastToPlotDto {
  forecast_date:     string;
  target:            string;
  target_end_date:   string;
  location:          string;
  type:              string;
  quantile:          string;
  value:             string;
  timezero:          string;
  model:             string;
  truth_data_source: string;
  shift_ECDC:        string;
  shift_JHU:         string;
  first_commit_date: string;
}